Refactoring: requirement.py -> node.py
