+++ strictdoc/backend/sdoc/models/fragment_from_file.py
@@ -1,4 +1,4 @@
-from strictdoc.backend.sdoc.models.sdoc_object import SDocObject
+from strictdoc.backend.sdoc.models.object import SDocObject
 from strictdoc.backend.sdoc.models.section import SectionContext
 from strictdoc.helpers.auto_described import auto_described
