+++ strictdoc/backend/sdoc/models/object.py
