+++ strictdoc/core/query_engine/query_object.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from strictdoc.backend.sdoc.models.document_grammar import GrammarElement
-from strictdoc.backend.sdoc.models.requirement import (
+from strictdoc.backend.sdoc.models.node import (
     Requirement,
     RequirementField,
 )
