+++ strictdoc/export/dot/document_dot_generator.py
@@ -7,7 +7,7 @@ from typing import List, Optional, Tuple, Union
 import graphviz
 
 from strictdoc.backend.sdoc.models.document import Document
-from strictdoc.backend.sdoc.models.requirement import Requirement
+from strictdoc.backend.sdoc.models.node import Requirement
 from strictdoc.backend.sdoc.models.section import Section
 from strictdoc.core.document_iterator import DocumentCachingIterator
 from strictdoc.core.project_config import ProjectConfig
