+++ strictdoc/export/html/renderers/markup_renderer.py
@@ -3,7 +3,7 @@ from typing import Optional, Type, Union
 from strictdoc.backend.sdoc.models.anchor import Anchor
 from strictdoc.backend.sdoc.models.document import Document
 from strictdoc.backend.sdoc.models.inline_link import InlineLink
-from strictdoc.backend.sdoc.models.requirement import (
+from strictdoc.backend.sdoc.models.node import (
     Requirement,
 )
 from strictdoc.backend.sdoc.models.section import FreeText, Section
