+++ strictdoc/server/routers/main_router.py
@@ -29,7 +29,7 @@ from strictdoc.backend.sdoc.models.document_grammar import (
     GrammarElement,
 )
 from strictdoc.backend.sdoc.models.document_view import ViewElement
-from strictdoc.backend.sdoc.models.requirement import (
+from strictdoc.backend.sdoc.models.node import (
     Requirement,
 )
 from strictdoc.backend.sdoc.models.section import Section
