+++ tests/unit/strictdoc/backend/sdoc/models/test_requirement.py
@@ -1,8 +1,8 @@
 from strictdoc.backend.sdoc.models.document import Document
 from strictdoc.backend.sdoc.models.document_config import DocumentConfig
 from strictdoc.backend.sdoc.models.document_grammar import DocumentGrammar
+from strictdoc.backend.sdoc.models.node import RequirementField
 from strictdoc.backend.sdoc.models.object_factory import SDocObjectFactory
-from strictdoc.backend.sdoc.models.requirement import RequirementField
 from strictdoc.backend.sdoc.models.type_system import GrammarElementFieldString
