+++ tests/unit/strictdoc/backend/sdoc/test_dsl_grammars.py
@@ -1,4 +1,4 @@
-from strictdoc.backend.sdoc.models.requirement import Requirement
+from strictdoc.backend.sdoc.models.node import Requirement
 from strictdoc.backend.sdoc.reader import SDReader
